Support for statistics windows as plugin (3d scatterplot) (#51)
* integrated 3d scatterplot

* removed printstacktrace

* initial commit on jzy3d scatterplot

* improved mouse event handling

* minor handling fixes

* added 3d scatterplot icon

* work on handling mouse events

* set version explicitly

* re-use functionality from color manipulation for 3d scatterplot

* use roi masks

* added projection scatter class

* added projection option

* avoid concurrent modification

* draw in correct order

* projection and line react to animation

* allow to reset scale

* enabled zoom all function

* update line immediately

* enabled save functionality

* allow to choose display level

* redesigned refreshactionenabler

* keep color band optional

* invert choice of level of detail

* use swingworker for updating chart

* improved picking performance

* only draw projections of data within axis bounds

* draw line only to axis plains

* inverted condition

* allow to only display data in axis bounds

* allow choosing between color band encoding and single color display

* preset transparency

* enabled axis range controls

* support editing of title label

* display picking result in dedicated label

* remove selection when not included in chart update

* only pick when shift is down

* deal with case that selection is empty

* display number of valid pixels

* avoid arrayindexoutofboundsexception

* cleanings

* allow setting non-numerical statistic measure

* moved roimaskselector to util

* renamed to colorformmodel

* made classes public

* removed 3d scatterplot

* revert

* removed dependency and repository

* made constructors public

* made public

* removed unneeded repo and dependency
