+++ snap-rcp/src/main/java/org/esa/snap/rcp/colormanip/ColorManipulationTopComponent.java
@@ -56,7 +56,7 @@ public class ColorManipulationTopComponent extends TopComponent implements HelpC
 
     public ColorManipulationTopComponent() {
         setName(Bundle.CTL_ColorManipulationTopComponent_ComponentName());
-        ColorManipulationForm cmf = new ColorManipulationForm(this, new FormModel());
+        ColorManipulationFormImpl cmf = new ColorManipulationFormImpl(this, new ColorFormModel());
         setLayout(new BorderLayout());
         add(cmf.getContentPanel(), BorderLayout.CENTER);
     }
