+++ snap-rcp/src/main/java/org/esa/snap/rcp/colormanip/Continuous3BandGraphicalForm.java
@@ -46,7 +46,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-class Continuous3BandGraphicalForm implements ColorManipulationChildForm {
+public class Continuous3BandGraphicalForm implements ColorManipulationChildForm {
 
     private final ColorManipulationForm parentForm;
     private final ImageInfoEditor2 imageInfoEditor;
@@ -146,7 +146,7 @@ class Continuous3BandGraphicalForm implements ColorManipulationChildForm {
     }
 
     @Override
-    public void handleFormShown(FormModel formModel) {
+    public void handleFormShown(ColorFormModel formModel) {
         RasterDataNode[] rasters = formModel.getRasters();
         initialChannelSources[0] = rasters[0];
         initialChannelSources[1] = rasters[1];
@@ -156,7 +156,7 @@ class Continuous3BandGraphicalForm implements ColorManipulationChildForm {
     }
 
     @Override
-    public void handleFormHidden(FormModel formModel) {
+    public void handleFormHidden(ColorFormModel formModel) {
         imageInfoEditor.setModel(null);
         channelSourcesList.clear();
         Arrays.fill(models, null);
@@ -165,7 +165,7 @@ class Continuous3BandGraphicalForm implements ColorManipulationChildForm {
     }
 
     @Override
-    public void updateFormModel(FormModel formModel) {
+    public void updateFormModel(ColorFormModel formModel) {
         RasterDataNode[] rasters = formModel.getRasters();
         currentChannelSources[0] = rasters[0];
         currentChannelSources[1] = rasters[1];
@@ -206,7 +206,7 @@ class Continuous3BandGraphicalForm implements ColorManipulationChildForm {
     }
 
     @Override
-    public void resetFormModel(FormModel formModel) {
+    public void resetFormModel(ColorFormModel formModel) {
         updateFormModel(formModel);
         imageInfoEditor.computeZoomOutToFullHistogramm();
     }
