+++ snap-rcp/src/main/java/org/esa/snap/rcp/statistics/DensityPlotPanel.java
@@ -135,9 +135,9 @@ class DensityPlotPanel extends ChartPagePanel {
 
     private void initActionEnablers() {
         RefreshActionEnabler roiMaskActionEnabler = new RefreshActionEnabler(refreshButton, PROPERTY_NAME_USE_ROI_MASK,
-                                                                             PROPERTY_NAME_ROI_MASK,
-                                                                             PROPERTY_NAME_X_PRODUCT, PROPERTY_NAME_Y_PRODUCT,
-                                                                             PROPERTY_NAME_X_BAND, PROPERTY_NAME_Y_BAND);
+                                                                             PROPERTY_NAME_ROI_MASK);
+        roiMaskActionEnabler.addProductBandEnablement(PROPERTY_NAME_X_PRODUCT, PROPERTY_NAME_X_BAND);
+        roiMaskActionEnabler.addProductBandEnablement(PROPERTY_NAME_Y_PRODUCT, PROPERTY_NAME_Y_BAND);
         bindingContext.addPropertyChangeListener(roiMaskActionEnabler);
         RefreshActionEnabler rangeControlActionEnabler = new RefreshActionEnabler(refreshButton, PROPERTY_NAME_MIN, PROPERTY_NAME_AUTO_MIN_MAX,
                                                                                   PROPERTY_NAME_MAX);
@@ -508,7 +508,6 @@ class DensityPlotPanel extends ChartPagePanel {
                     plot.getRangeAxis().setLabel(StatisticChartStyling.getAxisLabel(getRaster(Y_VAR), "Y", false));
                     toggleColorCheckBox.setEnabled(true);
                 } catch (InterruptedException | CancellationException e) {
-                    e.printStackTrace();
                     Dialogs.showMessage(CHART_TITLE,
                                             "Failed to compute scatter plot.\n" +
                                                     "Calculation canceled.",
@@ -516,7 +515,6 @@ class DensityPlotPanel extends ChartPagePanel {
                                         null
                     );
                 } catch (ExecutionException | IllegalArgumentException e) {
-                    e.printStackTrace();
                     Dialogs.showMessage(CHART_TITLE,
                                             "Failed to compute scatter plot.\n" +
                                                     "An error occurred:\n" +
